Microsoft SQL Server to work with my program

+++ RingSoft.DbLookup.EfCore/DbConstants.cs
@@ -6,7 +6,7 @@
         public const string StringColumnType = "nvarchar";
         public const string DecimalColumnType = "numeric";
         public const string DateColumnType = "datetime";
-        public const string ByteColumnType = "smallint";
+        public const string ByteColumnType = "tinyint";
         public const string BoolColumnType = "bit";
         public const string MemoColumnType = "ntext";
